pyproject.toml
[project]
name = "fastmcp-unified-template"
version = "0.5.0"
description = "MCP server template using FastMCP 3.x with FileSystemProvider discovery, standalone decorators, and OpenShift deployment"
requires-python = ">=3.11"
authors = [{name = "Your Name", email = "you@example.com"}]
readme = "README.md"
dependencies = [
  "fastmcp>=3.2.0",
  "pyyaml>=6.0.2",
  "python-dotenv>=1.1.1",
  "pytest>=8.4.1",
  "pytest-asyncio>=0.24.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[project.scripts]
fastmcp-unified = "src.main:main"

[tool.black]
line-length = 88

[tool.ruff]
line-length = 88

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
